Use APK sizes for progress reporting
to have more fine-grained progress reporting for e.g. slow downloads. The animation speed of the progress bar was reduced to give faster feedback for more common internet downloads as opposed to local file reads.

+++ app/src/main/java/org/calyxos/lupin/UiState.kt
@@ -16,7 +16,7 @@ sealed class UiState(
 
     class SelectionComplete(items: List<AppItem>) : UiState(items)
 
-    class InstallingApps(items: List<AppItem>, val done: Int, val total: Int) : UiState(items)
+    class InstallingApps(items: List<AppItem>, val done: Long, val total: Long) : UiState(items)
 
     class Done(items: List<AppItem>) : UiState(items)
